Extraced config.guess code from configuremake.py and used in rpackage.py
